pyproject.toml
[project]
name = "fastapi-todos"
version = "0.1.0"
description = "Endpoints Todos in fastapi"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "aiomysql>=0.2.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.13.3",
    "fastapi[all]>=0.119.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "slowapi>=0.1.9",
    "uvicorn[standard]>=0.38.0",
    "secure>=0.3.0",
    "redis>=5.0.0",
    "celery>=5.4,<6",
    "pyjwt>=2.8.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=5.0.0",
    "pre-commit>=3.5.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "black>=24.0.0",
    "types-redis>=4.6.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = [
    "-q",
    "--cov=app",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=xml",
    "--cov-report=html",
]
testpaths = ["tests"]

# ============================================================================
# Black - Code Formatter
# ============================================================================
[tool.black]
line-length = 120
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # Default excludes
  \.git
  | \.mypy_cache
  | \.ruff_cache
  | \.venv
  | \.eggs
  | \.tox
  | build
  | dist
  | __pycache__
)/
'''

# ============================================================================
# Ruff - Fast Python Linter
# ============================================================================
[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = [
    ".venv",
    "__pycache__",
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "W191",   # indentation contains tabs
    "B904",   # Allow raising exceptions without from e, for HTTPException
    "SIM105", # Allow try-except-pass (cleaner for optional cleanup)
    "ARG001", # Allow unused arguments (common in FastAPI)
    "E402",   # Allow module level imports not at top (needed for some configs)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/**/*.py" = [
    "ARG",    # Unused function args in tests
    "S101",   # Allow assert in tests
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# ============================================================================
# MyPy - Static Type Checker
# ============================================================================
[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
strict_optional = true
strict_equality = true
show_error_codes = true
show_column_numbers = true
pretty = true

# Exclude specific directories
exclude = [
    "^build/",
    "^dist/",
    "^\\.venv/",
    "^\\.mypy_cache/",
]

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "slowapi.*",
    "secure.*",
    "redis.*",
]
ignore_missing_imports = true

# ============================================================================
# Bandit - Security Linter
# ============================================================================
[tool.bandit]
exclude_dirs = [
    ".venv",
    "tests",
    "build",
    "dist",
]
skips = [
    "B101",  # Allow assert statements (used in tests)
    "B601",  # Allow shell=False (we use it safely)
]
